requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
dj-database-url==0.5.0
Django==3.0.2
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-crispy-forms==1.8.0
django-js-asset==1.2.2
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.3
soupsieve==1.9.4
sqlparse==0.3.0
whitenoise==4.1.4
